feat: Refactor CreateSingleItemModal and extract import files logic into a new component
