+++ src/components/Modals/CreateSingleItemModal/ImportStep/ImportStep.types.ts
@@ -1,0 +1,29 @@
+import { WearableCategory } from '@dcl/schemas'
+import { Item } from 'modules/item/types'
+import React from 'react'
+import { AcceptedFileProps } from '../CreateSingleItemModal.types'
+
+export type Props = {
+  category?: WearableCategory
+  metadata?: CreateSingleItemModalMetadata
+  title: string
+  wearablePreviewComponent?: React.ReactNode
+  isRepresentation?: boolean
+  onDropAccepted: (acceptedFileProps: AcceptedFileProps) => void
+  onDropRejected?: (files: File[]) => Promise<void>
+  onClose: () => void
+}
+
+export type StateData = {
+  id: string
+  error: string
+  isLoading: boolean
+}
+export type State = Partial<StateData>
+
+export type CreateSingleItemModalMetadata = {
+  collectionId?: string
+  item?: Item
+  addRepresentation?: boolean
+  changeItemFile?: boolean
+}
